test STORE-CV service
A lot of work going
The test doesn't pass yet since an adaptation is necessary in the
service to manage the changed JSON structure now define in the DTO
package!

So now DTO = class that here defines the INPUT as well as the OUTPUT
of the POST/GET methods for the cv service route.

Later on we can think of creating different INPUT/OUTPUT models.